better path-to-regexp diff logging
