+++ packages/node/package.json
@@ -29,6 +29,7 @@
     "@vercel/error-utils": "2.0.3",
     "@vercel/nft": "0.27.10",
     "@vercel/static-config": "3.0.0",
+    "@vercel-internals/path-to-regexp": "1.0.0",
     "async-listen": "3.0.0",
     "cjs-module-lexer": "1.2.3",
     "edge-runtime": "2.5.9",
@@ -36,8 +37,6 @@
     "esbuild": "0.14.47",
     "etag": "1.8.1",
     "node-fetch": "2.6.9",
-    "path-to-regexp": "6.2.1",
-    "path-to-regexp-updated": "npm:path-to-regexp@6.3.0",
     "ts-morph": "12.0.0",
     "ts-node": "10.9.1",
     "typescript": "4.9.5",
