+++ packages/remix/package.json
@@ -33,10 +33,9 @@
     "@types/node": "14.18.33",
     "@types/semver": "7.3.13",
     "@vercel/build-utils": "9.1.0",
+    "@vercel-internals/path-to-regexp": "1.0.0",
     "glob": "10.3.16",
     "jest-junit": "16.0.0",
-    "path-to-regexp": "6.2.1",
-    "path-to-regexp-updated": "npm:path-to-regexp@6.3.0",
     "semver": "7.5.2",
     "vitest": "2.0.1"
   },
